move around store files
