--- src/App.vue
@@ -1,17 +1,0 @@
-<template>
-  <div id="app">
-    <MainLayout>
-      <router-view />
-    </MainLayout>
-    <AccountModal />
-  </div>
-</template>
-
-<script>
-import AccountModal from '@/components/Modals/AccountModal';
-import MainLayout from '@/layouts/MainLayout';
-
-export default {
-  components: { AccountModal, MainLayout },
-};
-</script>
