+++ src/layouts/default.vue
@@ -1,15 +1,18 @@
 <template>
   <div id="main-layout">
     <MainMenu />
-    <slot></slot>
+    <nuxt></nuxt>
+    <AccountModal />
   </div>
 </template>
 
 <script>
+import AccountModal from '@/components/Modals/AccountModal';
 import MainMenu from '@/components/Navigation/MainMenu';
 
 export default {
   components: {
+    AccountModal,
     MainMenu,
   },
 };
